Make fxa.tests.utils importable without test dependencies.
